refactored textparser
added proper checks on the input file
moved the if/else for outfile so that it makes the file AFTER the file is successfully parsed. This had earlier generated the outfile even if the infile could not be parsed, leaving the outfile empty and worthless. There was also no rollback-functionality.